refactor(io): split io/ops.rs into per-section submodules
Convert `io/ops.rs` (1036 LOC) into `ops/` with four per-section
submodules:

- `config.rs` (78 LOC): `AuditLevel` enum, `SecureFileOpsConfig` struct,
  three preset constructors (`secure`, `development`, `performance`)
- `core.rs` (~530 LOC): `SecureFileOps` struct + all read/write/info/
  locked/permission methods (async + sync), plus the private observe
  helpers (`log_debug`, `log_info`, `log_error`, `start_timer`,
  `record_metric`)
- `builder.rs` (52 LOC): `SecureFileOpsBuilder` fluent configuration
- `tests.rs` (234 LOC): all `#[cfg(test)]` unit tests, with explicit
  `crate::primitives::io::file::FileMode` import (previously implicit
  via the parent module's `use` line)

`ops/mod.rs` (51 LOC) re-exports `AuditLevel`, `SecureFileOps`,
`SecureFileOpsBuilder`, `SecureFileOpsConfig` so existing call sites
through `io::SecureFileOps` etc. work unchanged.

`SecureFileOps.config` was bumped from private to `pub(super)` so the
sibling `tests.rs` module (under `ops/tests.rs`) can still assert on it.

Pure move-only refactor — no logic, signature, or external API changes.
Follows the precedent set by #317/#318/#331/#335/#341/#342/#343.

Part of #338 (file-length umbrella). 4 of 7 PRs.
