+++ crates/octarine/src/io/ops/core.rs
@@ -1,36 +1,8 @@
-//! SecureFileOps - Unified file operations with audit trails and metrics
+//! Core `SecureFileOps` implementation
 //!
-//! Provides a high-level API for file operations that automatically:
-//! - Logs all operations via observe
-//! - Records metrics (timing, counts, sizes)
-//! - Supports configurable audit levels
-//! - Validates file types via magic bytes
-//!
-//! # Design Philosophy
-//!
-//! SecureFileOps follows async-first design:
-//! - All primary operations are async to avoid blocking
-//! - Sync variants available with `_sync` suffix for use in sync contexts
-//! - All operations are logged for audit trails
-//! - Metrics are collected for monitoring
-//!
-//! # Examples
-//!
-//! ```ignore
-//! use octarine::io::SecureFileOps;
-//!
-//! // Create with default settings (audit enabled)
-//! let ops = SecureFileOps::new();
-//!
-//! // Read file with audit trail (async)
-//! let contents = ops.read_file("config.json").await?;
-//!
-//! // Write file atomically with audit trail (async)
-//! ops.write_file("output.txt", b"data".to_vec()).await?;
-//!
-//! // Sync variants for use in sync contexts
-//! let contents = ops.read_file_sync("config.json")?;
-//! ```
+//! Read/write/info/locked/permission operations plus the private observe-helper
+//! methods. The async/sync API symmetry and `Default` impl are preserved
+//! exactly as on the unsplit file.
 
 use std::io::{Read, Write};
 use std::path::Path;
@@ -43,82 +15,13 @@ use crate::primitives::io::file::{
     read_file as prim_read_file, read_file_sync as prim_read_file_sync,
 };
 
-use super::magic::{
+use crate::io::magic::{
     MagicResult, detect_file_type, detect_file_type_sync, validate_image, validate_image_sync,
     validate_not_dangerous, validate_not_dangerous_sync,
 };
 
-/// Audit level for file operations
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum AuditLevel {
-    /// No logging (not recommended for production)
-    Off,
-    /// Log errors only
-    Errors,
-    /// Log errors and warnings
-    Warnings,
-    /// Log all operations (default)
-    #[default]
-    Full,
-    /// Log all operations with debug details
-    Debug,
-}
-
-/// Configuration for SecureFileOps
-#[derive(Debug, Clone)]
-pub struct SecureFileOpsConfig {
-    /// Audit level for logging
-    pub audit_level: AuditLevel,
-    /// Whether to record metrics
-    pub metrics_enabled: bool,
-    /// Whether to validate magic bytes on read
-    pub validate_magic: bool,
-    /// Default write options
-    pub default_write_options: WriteOptions,
-}
-
-impl Default for SecureFileOpsConfig {
-    fn default() -> Self {
-        Self {
-            audit_level: AuditLevel::Full,
-            metrics_enabled: true,
-            validate_magic: false, // Opt-in for magic validation
-            default_write_options: WriteOptions::default(),
-        }
-    }
-}
-
-impl SecureFileOpsConfig {
-    /// Create config for high-security environments
-    pub fn secure() -> Self {
-        Self {
-            audit_level: AuditLevel::Full,
-            metrics_enabled: true,
-            validate_magic: true,
-            default_write_options: WriteOptions::for_secrets(),
-        }
-    }
-
-    /// Create config for development (verbose logging)
-    pub fn development() -> Self {
-        Self {
-            audit_level: AuditLevel::Debug,
-            metrics_enabled: true,
-            validate_magic: false,
-            default_write_options: WriteOptions::default(),
-        }
-    }
-
-    /// Create config for performance-critical scenarios
-    pub fn performance() -> Self {
-        Self {
-            audit_level: AuditLevel::Errors,
-            metrics_enabled: false,
-            validate_magic: false,
-            default_write_options: WriteOptions::default(),
-        }
-    }
-}
+use super::builder::SecureFileOpsBuilder;
+use super::config::{AuditLevel, SecureFileOpsConfig};
 
 /// Secure file operations with audit trails and metrics
 ///
@@ -143,7 +46,7 @@ impl SecureFileOpsConfig {
 /// ```
 #[derive(Debug, Clone)]
 pub struct SecureFileOps {
-    config: SecureFileOpsConfig,
+    pub(super) config: SecureFileOpsConfig,
 }
 
 impl Default for SecureFileOps {
@@ -751,286 +654,6 @@ impl SecureFileOps {
             // and the typed metrics API. For now, we use the timer for duration
             // and log the values via observe.
             // The observe event infrastructure provides audit trail.
-        }
-    }
-}
-
-/// Builder for SecureFileOps
-#[derive(Debug, Default)]
-pub struct SecureFileOpsBuilder {
-    config: SecureFileOpsConfig,
-}
-
-impl SecureFileOpsBuilder {
-    /// Create a new builder
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Set audit level
-    pub fn audit_level(mut self, level: AuditLevel) -> Self {
-        self.config.audit_level = level;
-        self
-    }
-
-    /// Enable or disable metrics
-    pub fn metrics(mut self, enabled: bool) -> Self {
-        self.config.metrics_enabled = enabled;
-        self
-    }
-
-    /// Enable or disable magic byte validation
-    pub fn validate_magic(mut self, enabled: bool) -> Self {
-        self.config.validate_magic = enabled;
-        self
-    }
-
-    /// Set default write options
-    pub fn default_write_options(mut self, options: WriteOptions) -> Self {
-        self.config.default_write_options = options;
-        self
-    }
-
-    /// Build the SecureFileOps
-    pub fn build(self) -> SecureFileOps {
-        SecureFileOps::with_config(self.config)
-    }
-}
-
-// ============================================================================
-// Tests
-// ============================================================================
-
-#[cfg(test)]
-mod tests {
-    #![allow(clippy::panic, clippy::expect_used)]
-
-    use super::*;
-    use crate::io::MagicFileType;
-    use tempfile::tempdir;
-
-    #[test]
-    fn test_secure_file_ops_new() {
-        let ops = SecureFileOps::new();
-        assert_eq!(ops.config.audit_level, AuditLevel::Full);
-        assert!(ops.config.metrics_enabled);
-    }
-
-    #[test]
-    fn test_write_and_read_sync() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("test.txt");
-
-        let ops = SecureFileOps::new();
-        ops.write_file_sync(&path, b"hello world").expect("write");
-
-        let contents = ops.read_file_sync(&path).expect("read");
-        assert_eq!(contents, b"hello world");
-    }
-
-    #[tokio::test]
-    async fn test_write_and_read_async() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("test_async.txt");
-
-        let ops = SecureFileOps::new();
-        ops.write_file(path.clone(), b"hello async".to_vec())
-            .await
-            .expect("write");
-
-        let contents = ops.read_file(path).await.expect("read");
-        assert_eq!(contents, b"hello async");
-    }
-
-    #[test]
-    fn test_write_and_read_string_sync() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("test.txt");
-
-        let ops = SecureFileOps::new();
-        ops.write_file_string_sync(&path, "hello string")
-            .expect("write");
-
-        let contents = ops.read_file_string_sync(&path).expect("read");
-        assert_eq!(contents, "hello string");
-    }
-
-    #[test]
-    fn test_exists_sync() {
-        let dir = tempdir().expect("create temp dir");
-        let existing = dir.path().join("exists.txt");
-        let missing = dir.path().join("missing.txt");
-
-        std::fs::write(&existing, b"test").expect("create file");
-
-        let ops = SecureFileOps::new();
-        assert!(ops.exists_sync(&existing));
-        assert!(!ops.exists_sync(&missing));
-    }
-
-    #[tokio::test]
-    async fn test_exists_async() {
-        let dir = tempdir().expect("create temp dir");
-        let existing = dir.path().join("exists_async.txt");
-        let missing = dir.path().join("missing_async.txt");
-
-        std::fs::write(&existing, b"test").expect("create file");
-
-        let ops = SecureFileOps::new();
-        assert!(ops.exists(existing).await.expect("exists"));
-        assert!(!ops.exists(missing).await.expect("exists"));
-    }
-
-    #[test]
-    fn test_file_size_sync() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("sized.txt");
-
-        std::fs::write(&path, b"12345").expect("create file");
-
-        let ops = SecureFileOps::new();
-        let size = ops.file_size_sync(&path).expect("get size");
-        assert_eq!(size, 5);
-    }
-
-    #[test]
-    fn test_builder() {
-        let ops = SecureFileOps::builder()
-            .audit_level(AuditLevel::Errors)
-            .metrics(false)
-            .validate_magic(true)
-            .build();
-
-        assert_eq!(ops.config.audit_level, AuditLevel::Errors);
-        assert!(!ops.config.metrics_enabled);
-        assert!(ops.config.validate_magic);
-    }
-
-    #[test]
-    fn test_config_presets() {
-        let secure = SecureFileOpsConfig::secure();
-        assert!(secure.validate_magic);
-        assert_eq!(secure.audit_level, AuditLevel::Full);
-
-        let dev = SecureFileOpsConfig::development();
-        assert_eq!(dev.audit_level, AuditLevel::Debug);
-
-        let perf = SecureFileOpsConfig::performance();
-        assert!(!perf.metrics_enabled);
-    }
-
-    #[test]
-    fn test_write_secrets_sync() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("secret.txt");
-
-        let ops = SecureFileOps::new();
-        ops.write_secrets_sync(&path, b"secret data")
-            .expect("write secrets");
-
-        let contents = ops.read_file_sync(&path).expect("read");
-        assert_eq!(contents, b"secret data");
-
-        // On Unix, check permissions
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let mode = std::fs::metadata(&path)
-                .expect("metadata")
-                .permissions()
-                .mode()
-                & 0o777;
-            assert_eq!(mode, 0o600);
-        }
-    }
-
-    #[test]
-    fn test_locked_read_write() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("locked.txt");
-
-        let ops = SecureFileOps::new();
-
-        // Write with lock
-        ops.write_locked(&path, b"locked data")
-            .expect("write locked");
-
-        // Read with lock
-        let contents = ops.read_locked(&path).expect("read locked");
-        assert_eq!(contents, b"locked data");
-    }
-
-    #[test]
-    fn test_detect_type_sync() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("test.png");
-
-        // Write PNG magic bytes
-        std::fs::write(&path, [0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A]).expect("write png");
-
-        let ops = SecureFileOps::new();
-        let result = ops.detect_type_sync(&path).expect("detect");
-        assert_eq!(result.file_type, Some(MagicFileType::Png));
-    }
-
-    #[test]
-    fn test_read_validated_image_sync() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("image.png");
-
-        // Write PNG magic bytes
-        std::fs::write(&path, [0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A]).expect("write png");
-
-        let ops = SecureFileOps::new();
-        let data = ops.read_validated_image_sync(&path).expect("read image");
-        assert_eq!(data.len(), 8);
-    }
-
-    #[test]
-    fn test_read_validated_image_rejects_non_image() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("fake.png");
-
-        // Write ELF magic (not an image)
-        std::fs::write(&path, [0x7F, 0x45, 0x4C, 0x46]).expect("write elf");
-
-        let ops = SecureFileOps::new();
-        let result = ops.read_validated_image_sync(&path);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_read_safe_rejects_dangerous() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("script.sh");
-
-        // Write shebang
-        std::fs::write(&path, b"#!/bin/bash\necho hello").expect("write script");
-
-        let ops = SecureFileOps::new();
-        let result = ops.read_safe_sync(&path);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_set_mode() {
-        let dir = tempdir().expect("create temp dir");
-        let path = dir.path().join("mode.txt");
-
-        std::fs::write(&path, b"test").expect("create file");
-
-        let ops = SecureFileOps::new();
-        ops.set_mode(&path, FileMode::PRIVATE).expect("set mode");
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let mode = std::fs::metadata(&path)
-                .expect("metadata")
-                .permissions()
-                .mode()
-                & 0o777;
-            assert_eq!(mode, 0o600);
         }
     }
 }
